backend/requirements.txt
typing
python-dotenv
langchain-groq
langchain
pydantic
fastapi
uvicorn